# Transform keyMirror to object literal