--- a/.babelrc
+++ b/.babelrc
@@ -22,6 +22,7 @@
     "transform-es2015-modules-commonjs",
     "transform-es3-member-expression-literals",
     "transform-es3-property-literals",
+    "./scripts/babel/key-mirror-to-literal",
     "./scripts/babel/transform-object-assign-require",
     "transform-react-jsx-source"
   ]