I was hoping that avoiding `keyMirror` during load would have a more dramatic effect, but I can't tell if it's noise or not. The p90 didn't move, and I'm always reluctant to sell something solely based on averages.

Results using the same [benchmark](https://gist.github.com/zertosh/e05e708262ee499626fc4998646c79a0) from #7493 with 100 iterations:

| engine | build | before: avg / p90 (ms) | after: avg / p90 (ms) |
| ------ | ----- | -------------: | -------------: |
| node | react.js (development) | 22.67 / 23 | 22.32 / 23 |
| node | react.js (production) | 20.41 / 21 | 19.83 / 21 |
| node | dist/react-with-addons.js | 26.79 / 29 | 26.20 / 27 |
| node | dist/react-with-addons.min.js | 18.17 / 19.5 | 18.03 / 19 |
| node | dist/react.js | 18.73 / 19 | 18.02 / 19 |
| node | dist/react.min.js | 12.87 / 13 | 12.48 / 13 |

It does result in some modest size improvements:

```
   raw     gz Compared to master @ e5f9ae27058052e72d6fb6d89c6ca93075e4089a
     =      = build/react-dom-fiber.js
     =      = build/react-dom-fiber.min.js
  -745    -92 build/react-dom-server.js
  +453     +7 build/react-dom-server.min.js
  -745    -98 build/react-dom.js
  +454     -5 build/react-dom.min.js
 -1322   -337 build/react-with-addons.js
  -166    -39 build/react-with-addons.min.js
 -1322   -339 build/react.js
  -166    -50 build/react.min.js
```

Given the low-impact of this PR, it won't hurt my feelings if it doesn't get merged.