@zpao 2016-08-15
Yea, this will break GCC. We still care about that for now. I think when we decide to stop caring about that, we'll get rid of the use and code style around it, making this pretty moot.

Since you're curious… the other side of this to look at is `keyOf`, which makes sure we get the same key to lookup for consistent crushing.