@zertosh 2016-08-15
Ah yes, this is not GCC safe 😞