@vjeux 2016-08-17
@zpao how hard would it be to write a test to make sure that we don't accidentally break GCC?