@zpao 2016-08-18
Not sure. We'd need to build some minimal app with React, compile it with GCC, and make sure it works. Won't be a jest test. Not sure the extent of the complexity that app would need to actually cover all the cases we might break.